+++ src/benchmark.cpp
@@ -37,7 +37,7 @@ std::vector<timeInt64> time_algorithm(Config cfg) {
                              cfg.max_position);
 
   std::vector<timeInt64> cycles_vec;
-  float *solution;
+  double *solution;
 
   // Run the actual algorithm and time it for n_iterations
   for (int rep = 0; rep < cfg.n_repetitions; ++rep) {
